Initial FrontEnd Commit
Configured Server Side Rendering as this would be beneficial for SEO although may not be needed for this project is a nice feature moving forward to further optimization. Also configured App Engine API as this allows the Angular App to communicate with Server, fetch Data and perform other server side operations as this project will not be a static application. We are fetching data from hacker news api. The rest of the code is the initial angular architecture from running project command